requirements.txt
# ═══════════════════════════════════════════════════════════════
# Pygmalion AI 完整项目依赖
# Python 3.8+
# ═══════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────
# 核心系统依赖
# ─────────────────────────────────────────────────────────────────

# AI 和深度学习
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0

# CLIP 模型支持（参考图功能）
# transformers 已包含 CLIP 支持，无需额外安装

# API 和网络
requests>=2.31.0
aiohttp>=3.8.0

# ─────────────────────────────────────────────────────────────────
# Web UI 依赖 (Flask + SocketIO)
# ─────────────────────────────────────────────────────────────────

# Web 框架
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1

# WebSocket 支持
websocket-client==1.6.2

# ─────────────────────────────────────────────────────────────────
# 配置和环境
# ─────────────────────────────────────────────────────────────────

python-dotenv==1.0.0

# ─────────────────────────────────────────────────────────────────
# 工具和工具
# ─────────────────────────────────────────────────────────────────

# 图片处理
Pillow>=9.5.0

# 日志和监控
loguru>=0.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# ─────────────────────────────────────────────────────────────────
# 开发依赖（可选）
# ─────────────────────────────────────────────────────────────────

# pytest>=7.0.0          # 单元测试
pytest>=7.0.0
# black>=23.0.0          # 代码格式化
# flake8>=6.0.0          # 代码检查
# mypy>=1.0.0            # 类型检查

# ═══════════════════════════════════════════════════════════════
# 安装命令：
#   pip install -r requirements.txt
#
# 仅安装核心依赖（不含 Web UI）：
#   pip install -r requirements.txt --ignore-installed Flask*
#
# 开发环境：
#   pip install -r requirements.txt
#   pip install pytest black flake8 mypy
# ═══════════════════════════════════════════════════════════════
